Added/Change - BW2: Vehicles & Infantry to have some winter ones too
What
=================
Added: All vehicles from the BWMod and Redd'n'Tanks to our 2BNB
Bundeswehr faction, so that everything is in one place for the mission
makers, and the crew are all using our custom loadouts.
Changed: Infantry to have Winter faction infantry also, as duplicates so
that they can crew the Winter vehicles. This way we don't need to
change all crew of vehicles if we ever do get Winter loadouts for the
infantry.

Issue ID: N/A